+++ cmd/podman/commands_remoteclient.go
@@ -1,0 +1,21 @@
+// +build remoteclient
+
+package main
+
+import "github.com/urfave/cli"
+
+func getAppCommands() []cli.Command {
+	return []cli.Command{}
+}
+
+func getImageSubCommands() []cli.Command {
+	return []cli.Command{}
+}
+
+func getContainerSubCommands() []cli.Command {
+	return []cli.Command{}
+}
+
+func getMainAppFlags() []cli.Flag {
+	return []cli.Flag{}
+}
